Move server-side classes into 'server' package.
