+++ core-java/src/main/java/org/spine3/server/EventBus.java
@@ -17,10 +17,12 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.spine3;
+package org.spine3.server;
 
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Multimap;
+import org.spine3.Event;
+import org.spine3.EventClass;
 import org.spine3.base.EventContext;
 import org.spine3.base.EventRecord;
 import org.spine3.error.MissingEventApplierException;
@@ -31,7 +33,7 @@ import java.util.Map;
 import java.util.concurrent.locks.ReadWriteLock;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 
-import static org.spine3.util.Methods.*;
+import static org.spine3.server.ServerMethods.*;
 
 /**
  * Manages incoming events to the appropriate registered handler
