+++ core-java/src/test/java/org/spine3/server/CommandBusTest.java
@@ -19,8 +19,6 @@
  */
 package org.spine3.server;
 
-import org.spine3.CommandDispatcher;
-
 /**
  * @author Mikhail Melnik
  */
