+++ grpc-sample/src/main/java/org/spine3/sample/server/BaseSampleServer.java
@@ -23,10 +23,10 @@ import io.grpc.ServerImpl;
 import io.grpc.stub.StreamObserver;
 import io.grpc.netty.NettyServerBuilder;
 import org.slf4j.Logger;
-import org.spine3.CommandStore;
-import org.spine3.Engine;
-import org.spine3.EventBus;
-import org.spine3.EventStore;
+import org.spine3.server.CommandStore;
+import org.spine3.server.Engine;
+import org.spine3.server.EventBus;
+import org.spine3.server.EventStore;
 import org.spine3.base.CommandRequest;
 import org.spine3.base.CommandResult;
 import org.spine3.base.CommandServiceGrpc;
